+++ src/main/java/com/treemarket/tree/mapper/AddressMapper.java
@@ -5,6 +5,6 @@ import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
 public interface AddressMapper {
-    Long getAddressId(AddressVO addressvo);
+    Long getAddressId(AddressVO addressVO);
 
 }
