+++ src/main/java/com/treemarket/tree/service/AddressService.java
@@ -3,5 +3,5 @@ package com.treemarket.tree.service;
 import com.treemarket.tree.domain.AddressVO;
 
 public interface AddressService {
-    Long getAddressId(AddressVO addressVO);
+    Long getAddressId(String inputAddress);
 }
